--- src/components/form-control/form-control.stories.mdx
@@ -1,40 +1,0 @@
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
-import { FormControl } from "..";
-
-<Meta
-  title="Forms/FormControl"
-  component={FormControl}
-  argTypes={{
-    error: {
-      control: "boolean",
-      description: "If true, the input will indicate an error. The prop defaults to the value (false)."
-    },
-  }}
-/>
-
-export const Template = (args) => <FormControl {...args} />;
-
-# Input Label
-
-A styled version of <a href="https://mui.com/base/react-form-control/" target="_blank">MUI's FormControlUnstyled</a>, so you can use any prop provided there.
-
-## Example
-
-```
-import React, { useState } from 'react';
-import { FormControl } from '@moderntribe/wme';
-
-export const MyFormControl = (props) => (
-  <FormControl>
-    ...
-  <FormControl>
-);
-
-export default MyFormControl;
-```
-
-## Canvas
-
-<Canvas>
-  <Story name="Base">{Template.bind({})}</Story>
-</Canvas>
