--- src/components/form-control/index.ts
@@ -1,1 +1,0 @@
-export { default } from './form-control';
