+++ src/components/form-field-control/index.ts
@@ -1,0 +1,1 @@
+export { default } from './form-field-control';
