+++ src/components/form-field/form-field.tsx
@@ -1,10 +1,7 @@
 import React, { PropsWithChildren, ReactNode } from "react";
 import { FormControlUnstyledProps } from "@mui/base";
 
-import InputError from "../input-error";
-import FormControl from "../form-control";
-import InputHelperText from "../input-helper-text";
-import FormFieldLabel from "../form-field-label";
+import { InputError, FormFieldControl, InputHelperText, FormFieldLabel } from "..";
 
 type FormFieldProps = FormControlUnstyledProps & {
   errorMessage?: string;
@@ -23,13 +20,13 @@ const FormField: React.FC<PropsWithChildren<FormFieldProps>> = ({
   label,
   ...props
 }) => (
-  <FormControl {...props}>
+  <FormFieldControl {...props}>
     {label && <FormFieldLabel>{label}</FormFieldLabel>}
     {field}
     {errorMessage && <InputError>{errorMessage}</InputError>}
     {helperText && <InputHelperText>{helperText}</InputHelperText>}
     {children}
-  </FormControl>
+  </FormFieldControl>
 );
 
 export default FormField;
